Use PhantomDataSendSync to avoid accidentally removing Send+Sync bounds
