+++ src/lib.rs
@@ -53,10 +53,7 @@ use codec::{
     DecodeAll,
     Encode,
 };
-use core::{
-    fmt::Debug,
-    marker::PhantomData,
-};
+use core::fmt::Debug;
 use derivative::Derivative;
 
 mod client;
@@ -191,7 +188,7 @@ pub enum Phase {
 )]
 pub struct WrapperKeepOpaque<T> {
     data: Vec<u8>,
-    _phantom: PhantomData<T>,
+    _phantom: PhantomDataSendSync<T>,
 }
 
 impl<T: Decode> WrapperKeepOpaque<T> {
@@ -216,7 +213,31 @@ impl<T: Decode> WrapperKeepOpaque<T> {
     pub fn from_encoded(data: Vec<u8>) -> Self {
         Self {
             data,
-            _phantom: PhantomData,
+            _phantom: PhantomDataSendSync::new(),
         }
     }
 }
+
+/// A version of [`std::marker::PhantomData`] that is also Send and Sync (which is fine
+/// because regardless of the generic param, it is always possible to Send + Sync this
+/// 0 size type).
+#[derive(Derivative, Encode, Decode, scale_info::TypeInfo)]
+#[derivative(
+    Clone(bound = ""),
+    PartialEq(bound = ""),
+    Debug(bound = ""),
+    Eq(bound = ""),
+    Default(bound = "")
+)]
+#[scale_info(skip_type_params(T))]
+#[doc(hidden)]
+pub struct PhantomDataSendSync<T>(core::marker::PhantomData<T>);
+
+impl<T> PhantomDataSendSync<T> {
+    pub(crate) fn new() -> Self {
+        Self(core::marker::PhantomData)
+    }
+}
+
+unsafe impl<T> Send for PhantomDataSendSync<T> {}
+unsafe impl<T> Sync for PhantomDataSendSync<T> {}
